+++ src/lib/assertNever.ts
@@ -1,5 +1,5 @@
-import { ErrorFactory } from "./errors";
+import { ErrorFactory } from './errors';
 
 export const assertNever = (x: never): never => {
-  throw ErrorFactory.newInternalError("Unexpected object: " + x);
-}
+  throw ErrorFactory.newInternalError('Unexpected object: ' + x);
+};