+++ src/practices/LanguageIndependent/EditorConfigIsPresentPractice.spec.ts
@@ -14,6 +14,7 @@ describe('EditorConfigIsPresentPractice', () => {
 
   afterEach(async () => {
     containerCtx.virtualFileSystemService.clearFileSystem();
+    containerCtx.practiceContext.fileInspector!.purgeCache();
   });
 
   it('Returns practicing if there is a .editorconfig', async () => {